(#23) added a list box for easier selection of sprites; next will be reordering

+++ editor/src/MerlotJsonArray.java
@@ -36,6 +36,6 @@ public class MerlotJsonArray {
 	}
 
 	public Double getDouble(int idx) {
-		return ((Double) this.array.get(idx)).doubleValue();
+		return (Double) this.array.get(idx);
 	}
 }
